refactor(Hooks): Move useIsInDocs hook and add tests
